synchronization

+++ app/src/main/java/com/example/todolist/feature/todolist/domain/use_case/task_list/DeleteTaskList.kt
@@ -1,5 +1,8 @@
 package com.example.todolist.feature.todolist.domain.use_case.task_list
 
+import com.example.todolist.common.Constants
+import com.example.todolist.feature.todolist.data.remote.dto.toTaskItemDto
+import com.example.todolist.feature.todolist.data.remote.dto.toTaskListDto
 import com.example.todolist.feature.todolist.domain.model.InvalidTaskListException
 import com.example.todolist.feature.todolist.domain.model.TaskList
 import com.example.todolist.feature.todolist.domain.repository.TaskListRepository
@@ -10,6 +13,21 @@ class DeleteTaskList(
 
     @Throws(InvalidTaskListException::class)
     suspend operator fun invoke(vararg taskList: TaskList) {
-        repository.deleteTaskList(*taskList)
+        val taskListDto = taskList.map { it.toTaskListDto(Constants.ANDROID_ID) }
+
+        val result = repository.deleteTaskListOnRemote(
+            taskListDto = *taskListDto.toTypedArray(),
+            Constants.ANDROID_ID
+        )
+
+        if(result.isExecuted) {
+            repository.deleteTaskList(*taskList)
+        } else {
+            val data = taskList.map {
+                val item = it.copy(needToBeDeleted = true)
+                item
+            }
+            repository.updateTaskList(*data.toTypedArray())
+        }
     }
 }
